created a BaseConcatDataset class. it implements a split function. MOABBDataset and TUHAbnormal inherit from BaseConcatDataset
